requirements.txt
torch==2.8.0+cu129
torchvision==0.23.0+cu129
numpy==2.3.3
matplotlib==3.10.6
pillow==11.3.0
tqdm==4.67.1
pandas==2.3.3 
streamlit==1.50.0